Move some utility functions from the webidl crate into the backend crate
