+++ crates/backend/src/lib.rs
@@ -12,3 +12,4 @@ extern crate wasm_bindgen_shared as shared;
 
 pub mod ast;
 mod codegen;
+pub mod util;
